# Pool events for final processing of combined metrics