Dependent on https://github.com/elastic/apm-data/pull/128

Benchmarking:

```
name                       old time/op    new time/op    delta
CombinedMetricsToBatch-10     202µs ± 0%     200µs ± 1%   -0.86%  (p=0.000 n=10+10)

name                       old alloc/op   new alloc/op   delta
CombinedMetricsToBatch-10    38.6kB ± 0%     4.2kB ± 0%  -89.15%  (p=0.000 n=8+9)

name                       old allocs/op  new allocs/op  delta
CombinedMetricsToBatch-10       308 ± 0%        95 ± 0%  -69.16%  (p=0.000 n=10+10)
```

Small gain in terms of `time/op`. For `time/op` it seems a lot of time is spent on `HDRHistogramRepresentation#Buckets`. I think it can be optimized for sparse representations, I have added a TODO.